+++ src/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_executable(yo main.cc)
-add_library(stuff STATIC lzcnt.c lzcnt.cc)
+add_library(stuff OBJECT lzcnt.cc)
 target_link_libraries(yo benchmark::benchmark stuff)
