+++ src/lzcnt.cc
@@ -1,23 +1,17 @@
-#include <x86intrin.h>
-
 #include "lzcnt.h"
 
 static bool go = lzcnt_available();
 
-extern "C" {
-__attribute__((target("no-lzcnt"))) int lzcnt_slow(uint64_t word) {
-  if (word == 0)
-    return 8 * sizeof(word);
-  return __builtin_clzll(word);
-}
-__attribute__((target("lzcnt"))) int lzcnt_fast(uint64_t word) {
+__attribute__((target("lzcnt"))) inline int lzcnt_fast(uint64_t word) {
   return __builtin_clzll(word);
 }
 
-int lzcnt_static_var(uint64_t word) {
-  if (go)
-    return lzcnt_fast(word);
+static int lzcnt_choose(uint64_t word) {
+  if (lzcnt_available())
+    lzcnt_pfunc = lzcnt_fast;
   else
-    return lzcnt_slow(word);
+    lzcnt_pfunc = lzcnt_slow;
+  return lzcnt_pfunc(word);
 }
-}
+
+int (*lzcnt_pfunc)(uint64_t word) = lzcnt_choose;
